医学信息处理/期末任务_肝脏图像分割/seg/requirements.txt
click==8.1.7
numpy==1.24.2
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==10.1.0
pydicom==2.4.4
torch==2.0.0+cu118
torchsummary==1.5.1
torchvision==0.15.1+cu118
tqdm==4.61.2
